+++ stellar-lend/Cargo.toml
@@ -47,7 +47,7 @@ members = [
   "contracts/vault-share",
   "contracts/auto-compound-vault",
   "contracts/pool-interfaces",
-  "contracts/yield-router",
+  "contracts/yield-router", "contracts/shared-events",
 ]
 exclude = [
   "fuzz",
