+++ stellar-lend/contracts/hello-world/Cargo.toml
@@ -9,6 +9,7 @@ crate-type = ["cdylib", "lib"]
 doctest = false
 
 [dependencies]
+shared-events = { path = "../shared-events" }
 soroban-sdk = { workspace = true }
 soroban-token-sdk = { workspace = true }
 stellarlend-amm = { path = "../amm" }
