+++ stellar-lend/contracts/hello-world/src/events.rs
@@ -1,4 +1,5 @@
 #![allow(unused_variables)]
+pub use shared_events::*;
 
 use soroban_sdk::{contractevent, Address, Env, String, Symbol, Vec};
 
@@ -8,456 +9,127 @@ use crate::types::{AssetStatus, ProposalType, VoteType};
 // Core Lending Events (Existing)
 // ============================================================================
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct DepositEvent {
-    pub user: Address,
-    pub asset: Option<Address>,
-    pub amount: i128,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct WithdrawalEvent {
-    pub user: Address,
-    pub asset: Option<Address>,
-    pub amount: i128,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct BorrowEvent {
-    pub user: Address,
-    pub asset: Option<Address>,
-    pub amount: i128,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct RepayEvent {
-    pub user: Address,
-    pub asset: Option<Address>,
-    pub amount: i128,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct LiquidationEvent {
-    pub liquidator: Address,
-    pub borrower: Address,
-    pub debt_asset: Option<Address>,
-    pub collateral_asset: Option<Address>,
-    pub debt_liquidated: i128,
-    pub collateral_seized: i128,
-    pub incentive_amount: i128,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct FlashLoanInitiatedEvent {
-    pub user: Address,
-    pub asset: Address,
-    pub amount: i128,
-    pub fee: i128,
-    pub callback: Address,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct FlashLoanRepaidEvent {
-    pub user: Address,
-    pub asset: Address,
-    pub amount: i128,
-    pub fee: i128,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct AdminActionEvent {
-    pub actor: Address,
-    pub action: Symbol,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct PriceUpdatedEvent {
-    pub actor: Address,
-    pub asset: Address,
-    pub price: i128,
-    pub decimals: u32,
-    pub oracle: Address,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct RiskParamsUpdatedEvent {
-    pub actor: Address,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct PauseStateChangedEvent {
-    pub actor: Address,
-    pub operation: Symbol,
-    pub paused: bool,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct PositionUpdatedEvent {
-    pub user: Address,
-    pub collateral: i128,
-    pub debt: i128,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct AnalyticsUpdatedEvent {
-    pub user: Address,
-    pub activity_type: String,
-    pub amount: i128,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct UserActivityTrackedEvent {
-    pub user: Address,
-    pub operation: Symbol,
-    pub amount: i128,
-    pub timestamp: u64,
-}
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 
 // ============================================================================
 // Asset-Specific Events (Carbon Asset Style)
 // ============================================================================
 
 #[allow(dead_code)]
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct MintEvent {
-    pub token_id: u32,
-    pub owner: Address,
-    pub project_id: String,
-    pub vintage_year: u64,
-    pub methodology_id: u32,
-}
+
 
 #[allow(dead_code)]
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct TransferEvent {
-    pub token_id: u32,
-    pub from: Address,
-    pub to: Address,
-}
+
 
 #[allow(dead_code)]
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct StatusChangeEvent {
-    pub token_id: u32,
-    pub old_status: Option<AssetStatus>,
-    pub new_status: AssetStatus,
-    pub changed_by: Address,
-}
+
 
 #[allow(dead_code)]
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct QualityScoreUpdatedEvent {
-    pub token_id: u32,
-    pub old_score: i128,
-    pub new_score: i128,
-    pub updated_by: Address,
-}
+
 
 #[allow(dead_code)]
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct ApproveEvent {
-    pub from: Address,
-    pub spender: Address,
-    pub amount: i128,
-    pub live_until_ledger: u32,
-}
+
 
 #[allow(dead_code)]
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct Sep41TransferEvent {
-    pub from: Address,
-    pub to: Address,
-    pub amount: i128,
-}
+
 
 #[allow(dead_code)]
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct Sep41BurnEvent {
-    pub from: Address,
-    pub amount: i128,
-}
+
 
 // ============================================================================
 // Governance Events
 // ============================================================================
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct GovernanceInitializedEvent {
-    pub admin: Address,
-    pub vote_token: Address,
-    pub voting_period: u64,
-    pub quorum_bps: u32,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct ProposalCreatedEvent {
-    pub proposal_id: u64,
-    pub proposer: Address,
-    pub proposal_type: ProposalType,
-    pub description: String,
-    pub start_time: u64,
-    pub end_time: u64,
-    pub created_at: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct VoteCastEvent {
-    pub proposal_id: u64,
-    pub voter: Address,
-    pub vote_type: VoteType,
-    pub voting_power: i128,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct VotePowerSnapshotTakenEvent {
-    pub proposal_id: u64,
-    pub voter: Address,
-    pub balance: i128,
-    pub snapshot_time: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct VoteLockedEvent {
-    pub voter: Address,
-    pub proposal_id: u64,
-    pub locked_amount: i128,
-    pub locked_until: u64,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct VoteDelegatedEvent {
-    pub delegator: Address,
-    pub delegatee: Address,
-    pub delegated_at: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct VoteDelegationRevokedEvent {
-    pub delegator: Address,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct SuspiciousGovActivityEvent {
-    pub proposal_id: u64,
-    pub voter: Address,
-    pub voter_power: i128,
-    pub total_supply_estimate: i128,
-    pub reason: Symbol,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct ProposalQueuedEvent {
-    pub proposal_id: u64,
-    pub execution_time: u64,
-    pub for_votes: i128,
-    pub against_votes: i128,
-    pub quorum_reached: bool,
-    pub threshold_met: bool,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct ProposalExecutedEvent {
-    pub proposal_id: u64,
-    pub executor: Address,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct ProposalFailedEvent {
-    pub proposal_id: u64,
-    pub for_votes: i128,
-    pub against_votes: i128,
-    pub quorum_reached: bool,
-    pub threshold_met: bool,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct ProposalCancelledEvent {
-    pub proposal_id: u64,
-    pub caller: Address,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct ProposalApprovedEvent {
-    pub proposal_id: u64,
-    pub approver: Address,
-    pub timestamp: u64,
-}
+
+
+
+
+
+
+
+
+
+
+
+
+
 
 #[allow(dead_code)]
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct GovernanceConfigUpdatedEvent {
-    pub admin: Address,
-    pub voting_period: Option<u64>,
-    pub execution_delay: Option<u64>,
-    pub quorum_bps: Option<u32>,
-    pub proposal_threshold: Option<i128>,
-    pub timestamp: u64,
-}
+
 
 // ============================================================================
 // Multisig Events
 // ============================================================================
 
 #[allow(dead_code)]
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct MultisigConfigUpdatedEvent {
-    pub admin: Address,
-    pub admins: Vec<Address>,
-    pub threshold: u32,
-    pub timestamp: u64,
-}
+
 
 // ============================================================================
 // Guardian & Recovery Events
 // ============================================================================
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct GuardianAddedEvent {
-    pub guardian: Address,
-    pub added_by: Address,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct GuardianRemovedEvent {
-    pub guardian: Address,
-    pub removed_by: Address,
-    pub timestamp: u64,
-}
+
+
 
 #[allow(dead_code)]
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct GuardianThresholdUpdatedEvent {
-    pub admin: Address,
-    pub old_threshold: u32,
-    pub new_threshold: u32,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct RecoveryStartedEvent {
-    pub old_admin: Address,
-    pub new_admin: Address,
-    pub initiator: Address,
-    pub expires_at: u64,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct RecoveryApprovedEvent {
-    pub approver: Address,
-    pub current_approvals: u32,
-    pub threshold: u32,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct RecoveryExecutedEvent {
-    pub old_admin: Address,
-    pub new_admin: Address,
-    pub executor: Address,
-    pub timestamp: u64,
-}
+
+
+
+
 
 // ============================================================================
 // Treasury Events
 // ============================================================================
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct TreasurySetEvent {
-    pub admin: Address,
-    pub treasury: Address,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct ReservesClaimedEvent {
-    pub admin: Address,
-    pub asset: Option<Address>,
-    pub recipient: Address,
-    pub amount: i128,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct FeeConfigUpdatedEvent {
-    pub admin: Address,
-    pub interest_fee_bps: i128,
-    pub liquidation_fee_bps: i128,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct LiquidationFeeCollectedEvent {
-    pub asset: Option<Address>,
-    pub fee_amount: i128,
-    pub timestamp: u64,
-}
+
+
+
+
 
 // ============================================================================
 // Core Lending Emitter Helpers
@@ -674,156 +346,51 @@ pub fn emit_liquidation_fee_collected(e: &Env, event: LiquidationFeeCollectedEve
 // Credit Scoring Events (from origin/main)
 // ============================================================================
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct CreditScoreUpdatedEvent {
-    pub user: Address,
-    pub old_score: i128,
-    pub new_score: i128,
-    pub reason: String,
-    pub timestamp: u64,
-}
+
 
 // ============================================================================
 // Timelock Events (from origin/main)
 // ============================================================================
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct TimelockQueuedEvent {
-    pub operation_id: u64,
-    pub proposer: Address,
-    pub ready_at: u64,
-    pub expires_at: u64,
-    pub delay: u64,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct TimelockExecutedEvent {
-    pub operation_id: u64,
-    pub executor: Address,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct TimelockCancelledEvent {
-    pub operation_id: u64,
-    pub caller: Address,
-    pub timestamp: u64,
-}
+
+
+
 
 // ============================================================================
 // Circuit Breaker Events (from origin/main)
 // ============================================================================
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct CircuitBreakerActivatedEvent {
-    pub activated_by: Address,
-    pub emergency_mode: bool,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct CircuitBreakerDeactivatedEvent {
-    pub deactivated_by: Address,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct WhitelistAddedEvent {
-    pub liquidator: Address,
-    pub added_by: Address,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct WhitelistRemovedEvent {
-    pub liquidator: Address,
-    pub removed_by: Address,
-    pub timestamp: u64,
-}
+
+
+
+
 
 // ============================================================================
 // Liquidation Queue Events (from HEAD)
 // ============================================================================
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct LiquidationQueuedEvent {
-    pub entry_id: u64,
-    pub borrower: Address,
-    pub liquidator: Address,
-    pub health_factor: i128,
-    pub priority_score: i128,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct LiquidationProcessedEvent {
-    pub entry_id: u64,
-    pub borrower: Address,
-    pub liquidator: Address,
-    pub executor: Address,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct LiquidationCancelledEvent {
-    pub entry_id: u64,
-    pub caller: Address,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct BatchLiquidationEvent {
-    pub liquidator: Address,
-    pub total_positions: u32,
-    pub successful: u32,
-    pub failed: u32,
-    pub total_debt_liquidated: i128,
-    pub total_collateral_seized: i128,
-    pub timestamp: u64,
-}
+
+
+
+
 
 // ============================================================================
 // Emergency Withdrawal Events
 // ============================================================================
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct EmergencyTriggeredEvent {
-    pub trigger: crate::types::EmergencyTrigger,
-    pub started_at: u64,
-    pub window_opens_at: u64,
-    pub window_closes_at: u64,
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct EmergencyCancelledEvent {
-    pub timestamp: u64,
-}
 
-#[contractevent]
-#[derive(Clone, Debug)]
-pub struct EmergencyWithdrawalEvent {
-    pub user: Address,
-    pub asset: Option<Address>,
-    pub amount: i128,
-    pub loss_share_bps: i128,
-    pub timestamp: u64,
-}
+
+
+
 
 pub fn emit_emergency_triggered(e: &Env, state: crate::types::EmergencyState) {
     EmergencyTriggeredEvent {
